Improved and added missing German translations
The translations are now all up-to-date with the original English texts. I set up my own copy of the site to test them, but didn't set up the backend so I couldn't test all of them (see if the length and "flow" fits). If it gets merged, I'll check the site and fix any eventual problems.

Thanks for making this project possible :b